+++ apps/web/components/auth/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Auth Components
+ *
+ * Export all authentication-related components
+ */
+
+export { AuthModal } from "./auth-modal";
+export { AuthModalWrapper } from "./auth-modal-wrapper";
+export { UserMenu } from "./user-menu";